docs: add styles to make link title visible on keyboard focus
This commit adds CSS to style link title attributes so that they are visible when you tab through links in the body copy of the docs. Though we don't currently require the use of title attributes on links, PR#36281 adds a recommendation to use them. Making the title of links visible helps give links context when tabbing through them.